requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.18.0
numpy>=1.20.0
matplotlib>=3.5.0
pillow>=8.4.0
wandb>=0.12.0
tqdm>=4.62.0 